# Conflicts:
#   api/routes/users.py

+++ api/routes/users.py
@@ -10,7 +10,6 @@ router = APIRouter()
 @router.get("/{id}", response_model=UserSchema)
 def userList(id: int, session: Session = Depends(get_db)):
     user = session.query(User).get(id)
-    print(user)
     if user is None:
         raise HTTPException(status_code=404, detail="User not found")
     response: UserSchema = user
@@ -25,18 +24,26 @@ def userList(session: Session = Depends(get_db)):
     return response
 
 
-@router.post("/")
+@router.post("/", status_code=status.HTTP_201_CREATED)
 def addUser(user: AddUser, session: Session = Depends(get_db)):
-    userObj = User(**user.dict())
-    session.add(userObj)
-    session.commit()
-    session.refresh(userObj)
-    return userObj
+    try:
+        password = auth.get_password_hash(user.password)
+        user.password = password
+        userObj = User(**user.dict())
+        session.add(userObj)
+        session.commit()
+        session.refresh(userObj)
+    except Exception:
+        raise HTTPException(status_code=400, detail="Email Already exists")
+    return "User Created Successfully"
+
 
 
 @router.put("/{id}")
 def updateUser(id: int, user: UserUpdateSchema, session: Session = Depends(get_db)):
     userObj = session.query(User).get(id)
+    if userObj is None:
+        raise HTTPException(status_code=404, detail="User not found")
     userObj.first_name = user.first_name
     userObj.last_name = user.last_name
     session.commit()
@@ -46,6 +53,8 @@ def updateUser(id: int, user: UserUpdateSchema, session: Session = Depends(get_d
 @router.delete("/{id}")
 def deleteUser(id: int, session: Session = Depends(get_db)):
     userObj = session.query(User).get(id)
+    if userObj is None:
+        raise HTTPException(status_code=404, detail="User not found")
     session.delete(userObj)
     session.commit()
     session.close
